Make the import use a new RunningIcon component
